requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dateparser==1.2.1
distro==1.9.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
h3==4.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.3
openai==0.28.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
timezonefinder==6.5.8
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
yarl==1.18.3
